Split generic specialization integration test
